RaspPi Ver 1.0.8 End Node Ver 1.0.4
End Node
- Fix ACS712 - IADC bug
- Add Watchdog
RaspPi
- Fix struct failure

+++ Raspberryi Pi/mqtt.h
@@ -52,7 +52,7 @@ typedef struct
 	uint32_t timestamp;
 	uint8_t on_off;
 	uint8_t dimming;
-	uint8_t current_sensor;
+	uint16_t current_sensor;
 }MQTT_LED_Data_t;
 
 typedef struct
